Clarify bitmap estimator API
