+++ duckreg/__init__.py
@@ -30,14 +30,11 @@ from .regularized import DuckRidge
 from .bitpacking import (
     BitmapMundlak,
     BitmapMundlakEventStudy,
-    BitmapPanel,
-    LongPanelBitPacker,
 )
 
 __all__ = [
     "BitmapMundlak",
     "BitmapMundlakEventStudy",
-    "BitmapPanel",
     "DBDML",
     "DBDoubleDemeaning",
     "DBLogisticRegression",
@@ -59,7 +56,6 @@ __all__ = [
     "DuckReg",
     "DuckRegression",
     "DuckRidge",
-    "LongPanelBitPacker",
     "ridge_closed_form",
     "ridge_closed_form_batch",
     "wls",
